Moving main functions to a new script (find_and_analyse_changelogs)
Keeping all the old ones for the moment at least, but they're so needlessly messy.

Moved main changelog from local diff dicts to a variant of the sphinx_changelog_gen from blender source, it's neater, quicker and better serves the purpose (unsurprisingly). Removed the changelog generation section as it's not required here, but included a link to the original